관리자페이지 회원정보 - 상세(기본정보, 진행현황) 구현

+++ HealthPlanner/src/main/java/com/health/domain/MemberStateDTO.java
@@ -31,7 +31,7 @@ public class MemberStateDTO {
 	private int mbrWeight;
 
 	/** 체지방률 */
-	private int memberFatper;
+	private double memberFatper;
 
 	/** 종합 신체 등급 */
 	private String averageRank;
